page refactor

+++ core/domain/content/content.go
@@ -13,24 +13,24 @@ import "github.com/ixre/go2o/core/domain/interface/content"
 var _ content.IContent = new(Content)
 
 type Content struct {
-	contentRepo    content.IContentRepo
+	contentRepo    content.IArchiveRepo
 	userId         int64
 	articleManager content.IArticleManager
 }
 
-func NewContent(userId int64, rep content.IContentRepo) content.IContent {
+func NewContent(userId int64, rep content.IArchiveRepo) content.IContent {
 	return &Content{
 		contentRepo: rep,
 		userId:      userId,
 	}
 }
 
-// 获取聚合根编号
+// GetAggregateRootId 获取聚合根编号
 func (c *Content) GetAggregateRootId() int {
 	return int(c.userId)
 }
 
-// 文章服务
+// ArticleManager 文章服务
 func (c *Content) ArticleManager() content.IArticleManager {
 	if c.articleManager == nil {
 		c.articleManager = newArticleManagerImpl(c.userId, c.contentRepo)
@@ -38,12 +38,12 @@ func (c *Content) ArticleManager() content.IArticleManager {
 	return c.articleManager
 }
 
-// 创建页面
+// CreatePage 创建页面
 func (c *Content) CreatePage(v *content.Page) content.IPage {
 	return newPage(int32(c.GetAggregateRootId()), c.contentRepo, v)
 }
 
-// 获取页面
+// GetPage 获取页面
 func (c *Content) GetPage(id int32) content.IPage {
 	v := c.contentRepo.GetPageById(int32(c.GetAggregateRootId()), id)
 	if v != nil {
@@ -52,16 +52,23 @@ func (c *Content) GetPage(id int32) content.IPage {
 	return nil
 }
 
-// 根据字符串标识获取页面
-func (c *Content) GetPageByStringIndent(indent string) content.IPage {
-	v := c.contentRepo.GetPageByStringIndent(int32(c.GetAggregateRootId()), indent)
+// GetPageByCode 根据字符串标识获取页面
+func (c *Content) GetPageByCode(indent string) content.IPage {
+	v := c.contentRepo.GetPageByCode(c.GetAggregateRootId(), indent)
 	if v != nil {
 		return c.CreatePage(v)
 	}
 	return nil
 }
 
-// 删除页面
+// DeletePage 删除页面
 func (c *Content) DeletePage(id int32) error {
+	ip := c.GetPage(id)
+	if ip == nil{
+		return content.ErrNoSuchPage
+	}
+	if ip.GetValue().Flag & content.FlagInternal == content.FlagInternal{
+		return content.ErrInternalPage
+	}
 	return c.contentRepo.DeletePage(int32(c.GetAggregateRootId()), id)
 }
